refactor(adapt): split pinocchio_profile.rs into per-concern submodules
Split the 3,756-line adapt/pinocchio_profile.rs into a directory module with
per-concern submodules, mirroring the spec/chumsky_{adapter,parser} splits.
Move-only: no behavior or generated-output changes.

- types.rs    profile data model (Pinocchio* structs + merge/lookup impls)
- abi.rs      ABI-schema parsing and fold-into-profile helpers
- ast_util.rs low-level syn-expr walking and extractors
- infer.rs    AST/source-text inference over handler bodies
- mod.rs      facade: shared imports, inference entry points, re-exports
- tests.rs    the #[cfg(test)] inference tests

The directory rename keeps crate::adapt::pinocchio_profile (and the root
re-export crate::pinocchio_profile) intact; the 12 Pinocchio* structs and
infer_from_context stay reachable as crate::pinocchio_profile::<name> for
codegen::kani_impl. Cross-submodule items are pub(super); the facade scopes
non-pub(crate) globs to pub(in crate::adapt::pinocchio_profile).
